+++ benches/encoding.rs
@@ -22,8 +22,9 @@ use rand_xorshift::XorShiftRng;
 
 use exonum_crypto::{self, CryptoHash, Hash};
 use exonum_merkledb::{
+    impl_object_hash_for_binary_value,
     proof_map_index::{BranchNode, ProofPath},
-    BinaryKey, BinaryValue, UniqueHash,
+    BinaryKey, BinaryValue, ObjectHash,
 };
 
 const CHUNK_SIZE: usize = 64;
@@ -62,11 +63,8 @@ impl BinaryValue for SimpleData {
     }
 }
 
-impl CryptoHash for SimpleData {
-    fn hash(&self) -> Hash {
-        exonum_crypto::hash(&self.into_bytes())
-    }
-}
+impl_object_hash_for_binary_value! { SimpleData }
+impl_object_hash_for_binary_value! { CursorData }
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 struct CursorData {
@@ -103,12 +101,6 @@ impl BinaryValue for CursorData {
     }
 }
 
-impl CryptoHash for CursorData {
-    fn hash(&self) -> Hash {
-        exonum_crypto::hash(&self.into_bytes())
-    }
-}
-
 fn gen_bytes_data() -> Vec<u8> {
     let mut rng = XorShiftRng::from_seed(SEED);
     let mut v = vec![0; CHUNK_SIZE];
@@ -141,7 +133,7 @@ fn gen_branch_node_data() -> BranchNode {
 fn bench_binary_value<F, V>(c: &mut Criterion, name: &str, f: F)
 where
     F: Fn() -> V + 'static + Clone + Copy,
-    V: BinaryValue + UniqueHash + PartialEq + Debug,
+    V: BinaryValue + ObjectHash + PartialEq + Debug,
 {
     // Checks that binary value is correct.
     let val = f();
@@ -176,7 +168,7 @@ where
     c.bench_function(
         &format!("encoding/{}/hash", name),
         move |b: &mut Bencher| {
-            b.iter_with_setup(f, |data| black_box(data.hash()));
+            b.iter_with_setup(f, |data| black_box(data.object_hash()));
         },
     );
 }