+++ src/macros.rs
@@ -35,3 +35,14 @@ macro_rules! concat_keys {
         buf
     });
 }
+
+#[macro_export]
+macro_rules! impl_object_hash_for_binary_value {
+    ($type:ty) => {
+        impl ObjectHash for $type {
+            fn object_hash(&self) -> Hash {
+                exonum_crypto::hash(&self.to_bytes())
+            }
+        }
+    };
+}