+++ src/proof_list_index/tests.rs
@@ -23,7 +23,7 @@ use serde_json::{from_str, to_string};
 use exonum_crypto::Hash;
 
 use crate::{
-    hash::{HashTag, UniqueHash},
+    hash::{HashTag, ObjectHash},
     proof_list_index::{
         ListProof::{
             self, {Full, Leaf, Left, Right},
@@ -142,27 +142,27 @@ fn test_list_index_proof() {
     let h22 = HashTag::hash_single_node(&h2);
     let h012 = HashTag::hash_node(&h01, &h22);
 
-    assert_eq!(index.root_hash(), HashTag::empty_list_hash());
+    assert_eq!(index.object_hash(), HashTag::empty_list_hash());
 
     index.push(2u64);
 
-    assert_eq!(index.root_hash(), HashTag::hash_list_node(1, h0));
+    assert_eq!(index.object_hash(), HashTag::hash_list_node(1, h0));
     assert_eq!(index.get_proof(0), Leaf(2));
     assert_eq!(
         index
             .get_proof(0)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(0, &2)]
     );
 
     index.push(4u64);
-    assert_eq!(index.root_hash(), HashTag::hash_list_node(2, h01));
+    assert_eq!(index.object_hash(), HashTag::hash_list_node(2, h01));
     assert_eq!(index.get_proof(0), Left(Box::new(Leaf(2)), Some(h1)));
     assert_eq!(
         index
             .get_proof(0)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(0, &2)]
     );
@@ -170,7 +170,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_proof(1)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(1, &4)]
     );
@@ -182,13 +182,13 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_range_proof(0..2)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(0, &2), (1, &4)]
     );
 
     index.push(6u64);
-    assert_eq!(index.root_hash(), HashTag::hash_list_node(3, h012));
+    assert_eq!(index.object_hash(), HashTag::hash_list_node(3, h012));
     assert_eq!(
         index.get_proof(0),
         Left(Box::new(Left(Box::new(Leaf(2)), Some(h1))), Some(h22))
@@ -196,7 +196,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_proof(0)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(0, &2)]
     );
@@ -207,7 +207,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_proof(1)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(1, &4)]
     );
@@ -218,7 +218,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_proof(2)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(2, &6)]
     );
@@ -233,7 +233,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_range_proof(0..2)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(0, &2), (1, &4)]
     );
@@ -248,7 +248,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_range_proof(1..3)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(1, &4), (2, &6)]
     );
@@ -263,7 +263,7 @@ fn test_list_index_proof() {
     assert_eq!(
         index
             .get_range_proof(0..3)
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap(),
         [(0, &2), (1, &4), (2, &6)]
     );
@@ -281,7 +281,7 @@ fn test_randomly_generate_proofs() {
         index.push(value.clone());
     }
     index.get(0);
-    let table_merkle_root = index.root_hash();
+    let table_merkle_root = index.object_hash();
 
     for _ in 0..50 {
         let start_range = rng.gen_range(0, num_values);
@@ -328,7 +328,7 @@ fn test_index_and_proof_roots() {
     let db = TemporaryDB::default();
     let fork = db.fork();
     let mut index = ProofListIndex::new(IDX_NAME, &fork);
-    assert_eq!(index.root_hash(), HashTag::empty_list_hash());
+    assert_eq!(index.object_hash(), HashTag::empty_list_hash());
 
     let h1 = hash_leaf_node(&[1, 2]);
     let h2 = hash_leaf_node(&[2, 3]);
@@ -376,11 +376,11 @@ fn test_index_and_proof_roots() {
     for (inserted, exp_root, proof_ind) in expected_hash_comb {
         index.push(inserted);
 
-        assert_eq!(index.root_hash(), exp_root);
+        assert_eq!(index.object_hash(), exp_root);
         let range_proof = index.get_range_proof(proof_ind..proof_ind + 1);
         assert_eq!(
             range_proof
-                .validate(index.root_hash(), index.len())
+                .validate(index.object_hash(), index.len())
                 .unwrap()
                 .len(),
             1
@@ -391,7 +391,7 @@ fn test_index_and_proof_roots() {
         let range_proof = index.get_range_proof(0..proof_ind + 1);
         assert_eq!(
             range_proof
-                .validate(index.root_hash(), index.len())
+                .validate(index.object_hash(), index.len())
                 .unwrap()
                 .len(),
             (proof_ind + 1) as usize
@@ -402,7 +402,7 @@ fn test_index_and_proof_roots() {
         let range_proof = index.get_range_proof(0..1);
         assert_eq!(
             range_proof
-                .validate(index.root_hash(), index.len())
+                .validate(index.object_hash(), index.len())
                 .unwrap()
                 .len(),
             1
@@ -414,7 +414,7 @@ fn test_index_and_proof_roots() {
 
     let range_proof = index.get_range_proof(0..8);
     let (indices, val_refs): (Vec<_>, Vec<_>) = range_proof
-        .validate(index.root_hash(), index.len())
+        .validate(index.object_hash(), index.len())
         .unwrap()
         .into_iter()
         .unzip();
@@ -437,7 +437,7 @@ fn test_index_and_proof_roots() {
     let mut range_proof = index.get_range_proof(3..5);
     assert_eq!(
         range_proof
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap()
             .len(),
         2
@@ -445,7 +445,7 @@ fn test_index_and_proof_roots() {
     range_proof = index.get_range_proof(2..6);
     assert_eq!(
         range_proof
-            .validate(index.root_hash(), index.len())
+            .validate(index.object_hash(), index.len())
             .unwrap()
             .len(),
         4
@@ -460,7 +460,7 @@ fn test_proof_illegal_lower_bound() {
     let mut index = ProofListIndex::new(IDX_NAME, &fork);
     let proof = index.get_range_proof(0..1);
 
-    assert_proof_of_absence(proof, index.root_hash(), index.len());
+    assert_proof_of_absence(proof, index.object_hash(), index.len());
     index.push(vec![1]);
 }
 
@@ -473,7 +473,7 @@ fn test_proof_illegal_bound_empty() {
         index.push(vec![i]);
     }
     let proof = index.get_range_proof(8..9);
-    assert_proof_of_absence(proof, index.root_hash(), index.len());
+    assert_proof_of_absence(proof, index.object_hash(), index.len());
 }
 
 #[test]
@@ -493,7 +493,7 @@ fn test_proof_structure() {
     let db = TemporaryDB::default();
     let fork = db.fork();
     let mut index = ProofListIndex::new(IDX_NAME, &fork);
-    assert_eq!(index.root_hash(), HashTag::empty_list_hash());
+    assert_eq!(index.object_hash(), HashTag::empty_list_hash());
 
     // spell-checker:ignore upup
 
@@ -515,7 +515,7 @@ fn test_proof_structure() {
 
     let list_hash = HashTag::hash_list_node(index.len(), h12345);
 
-    assert_eq!(index.root_hash(), list_hash);
+    assert_eq!(index.object_hash(), list_hash);
     let range_proof = index.get_range_proof(4..5);
 
     assert_eq!(
@@ -556,10 +556,10 @@ fn test_simple_merkle_root() {
     let mut index = ProofListIndex::new(IDX_NAME, &fork);
     assert_eq!(index.get(0), None);
     index.push(vec![1]);
-    assert_eq!(index.root_hash(), h1);
+    assert_eq!(index.object_hash(), h1);
 
     index.set(0, vec![2]);
-    assert_eq!(index.root_hash(), h2);
+    assert_eq!(index.object_hash(), h2);
 }
 
 #[test]
@@ -587,7 +587,7 @@ fn test_same_merkle_root() {
     i2.push(vec![5]);
     i2.push(vec![1]);
 
-    assert_eq!(i1.root_hash(), i2.root_hash());
+    assert_eq!(i1.object_hash(), i2.object_hash());
 }
 
 #[derive(Serialize)]
@@ -600,11 +600,13 @@ struct ProofInfo<'a, V: Serialize + 'a> {
 }
 
 mod root_hash_tests {
-    use crate::{hash::HashTag, proof_list_index::ProofListIndex, Database, TemporaryDB};
+    use crate::{
+        hash::HashTag, proof_list_index::ProofListIndex, Database, ObjectHash, TemporaryDB,
+    };
     use exonum_crypto::{self, Hash};
 
-    /// Cross-verify `root_hash()` with `ProofListIndex` against expected root hash value.
-    fn assert_root_hash_correct(hashes: &[Hash]) {
+    /// Cross-verify `object_hash()` with `ProofListIndex` against expected root hash value.
+    fn assert_object_hash_correct(hashes: &[Hash]) {
         let root_actual = HashTag::hash_list(hashes);
         let root_index = proof_list_index_root(hashes);
         assert_eq!(root_actual, root_index);
@@ -615,7 +617,7 @@ mod root_hash_tests {
         let fork = db.fork();
         let mut index = ProofListIndex::new("merkle_root", &fork);
         index.extend(hashes.iter().cloned());
-        index.root_hash()
+        index.object_hash()
     }
 
     fn hash_list(bytes: &[&[u8]]) -> Vec<Hash> {
@@ -626,23 +628,23 @@ mod root_hash_tests {
     }
 
     #[test]
-    fn root_hash_single() {
-        assert_root_hash_correct(&hash_list(&[b"1"]));
+    fn object_hash_single() {
+        assert_object_hash_correct(&hash_list(&[b"1"]));
     }
 
     #[test]
-    fn root_hash_even() {
-        assert_root_hash_correct(&hash_list(&[b"1", b"2", b"3", b"4"]));
+    fn object_hash_even() {
+        assert_object_hash_correct(&hash_list(&[b"1", b"2", b"3", b"4"]));
     }
 
     #[test]
-    fn root_hash_odd() {
-        assert_root_hash_correct(&hash_list(&[b"1", b"2", b"3", b"4", b"5"]));
+    fn object_hash_odd() {
+        assert_object_hash_correct(&hash_list(&[b"1", b"2", b"3", b"4", b"5"]));
     }
 
     #[test]
-    fn root_hash_empty() {
-        assert_root_hash_correct(&hash_list(&[]));
+    fn object_hash_empty() {
+        assert_object_hash_correct(&hash_list(&[]));
     }
 }
 
@@ -656,11 +658,11 @@ fn proof_of_absence() {
         list.push(vec![i]);
     }
 
-    let root_hash =
+    let object_hash =
         Hash::from_hex("5ba859b4d1799cb27ece9db8f7a76a50fc713a5d9d22f753eca42172996a88f9").unwrap();
 
     let non_existed_index = 6_u64;
-    let expected_hash = HashTag::hash_list_node(list.len(), root_hash);
+    let expected_hash = HashTag::hash_list_node(list.len(), object_hash);
 
     let proof = list.get_proof(non_existed_index);
     assert_proof_of_absence(proof, expected_hash, list.len());
@@ -679,11 +681,11 @@ fn proof_of_absence_range() {
         list.push(vec![i]);
     }
 
-    let root_hash =
+    let object_hash =
         Hash::from_hex("5ba859b4d1799cb27ece9db8f7a76a50fc713a5d9d22f753eca42172996a88f9").unwrap();
 
     let non_existed_index = 6_u64;
-    let expected_hash = HashTag::hash_list_node(list.len(), root_hash);
+    let expected_hash = HashTag::hash_list_node(list.len(), object_hash);
 
     let proof = list.get_range_proof(2..);
     assert!(proof.validate(expected_hash, list.len()).is_ok());
@@ -692,7 +694,7 @@ fn proof_of_absence_range() {
     assert_proof_of_absence(proof, expected_hash, list.len());
 }
 
-fn assert_proof_of_absence<V: BinaryValue + UniqueHash + Debug>(
+fn assert_proof_of_absence<V: BinaryValue + ObjectHash + Debug>(
     proof: ListProof<V>,
     expected_hash: Hash,
     len: u64,