+++ src/proof_map_index/mod.rs
@@ -31,9 +31,8 @@ use self::{
 };
 use crate::{
     views::{IndexAccess, IndexBuilder, IndexType, Iter as ViewIter, View},
-    BinaryKey, BinaryValue, UniqueHash,
+    BinaryKey, BinaryValue, HashTag, ObjectHash,
 };
-use crate::hash::ObjectHash;
 
 mod key;
 mod node;
@@ -132,7 +131,7 @@ impl<T, K, V> ProofMapIndex<T, K, V>
 where
     T: IndexAccess,
     K: BinaryKey + ObjectHash,
-    V: BinaryValue + UniqueHash,
+    V: BinaryValue + ObjectHash,
 {
     /// Creates a new index representation based on the name and storage view.
     ///
@@ -251,37 +250,9 @@ where
     pub(crate) fn merkle_root(&self) -> Hash {
         match self.get_root_node() {
             Some((path, Node::Leaf(hash))) => HashTag::hash_single_entry_map(&path, &hash),
-            Some((_, Node::Branch(branch))) => branch.hash(),
+            Some((_, Node::Branch(branch))) => branch.object_hash(),
             None => Hash::zero(),
         }
-    }
-
-    /// Returns the hash of the proof map object. See [`HashTag::hash_map_node`].
-    /// For hash of the empty map see [`HashTag::empty_map_hash`].
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use exonum_merkledb::{TemporaryDB, Database, ProofMapIndex, HashTag};
-    /// use exonum_crypto::Hash;
-    ///
-    /// let db = TemporaryDB::new();
-    /// let name = "name";
-    /// let fork = db.fork();
-    /// let mut index = ProofMapIndex::new(name, &fork);
-    ///
-    /// let default_hash = index.object_hash();
-    /// assert_eq!(HashTag::empty_map_hash(), default_hash);
-    ///
-    /// index.put(&default_hash, 100);
-    /// let hash = index.object_hash();
-    /// assert_ne!(hash, default_hash);
-    /// ```
-    ///
-    /// [`HashTag::hash_map_node`]: ../enum.HashTag.html#method.hash_map_node
-    /// [`HashTag::empty_map_hash`]: ../enum.HashTag.html#method.empty_map_hash
-    pub fn object_hash(&self) -> Hash {
-        HashTag::hash_map_node(self.merkle_root())
     }
 
     /// Returns a value corresponding to the key.
@@ -537,7 +508,7 @@ where
 
     fn insert_leaf(&mut self, proof_path: &ProofPath, key: &K, value: V) -> Hash {
         debug_assert!(proof_path.is_leaf());
-        let hash = value.hash();
+        let hash = value.object_hash();
         self.base.put(proof_path, hash);
         self.base.put(&key.to_value_path(), value);
         hash
@@ -586,7 +557,7 @@ where
                             }
                             None => branch.set_child_hash(proof_path.bit(i), &h),
                         };
-                        let hash = branch.hash();
+                        let hash = branch.object_hash();
                         self.base.put(&child_path, branch);
                         (None, hash)
                     }
@@ -606,7 +577,7 @@ where
                 &parent.child_hash(proof_path.bit(0)),
             );
 
-            let hash = new_branch.hash();
+            let hash = new_branch.object_hash();
             self.base.put(&proof_path.prefix(i), new_branch);
             (Some(i), hash)
         }
@@ -643,14 +614,14 @@ where
                         RemoveAction::Branch((key, hash)) => {
                             let new_child_path = key.start_from(suffix_path.start());
                             branch.set_child(suffix_path.bit(0), &new_child_path, &hash);
-                            let h = branch.hash();
+                            let h = branch.object_hash();
 
                             self.base.put(&child_path, branch);
                             return RemoveAction::UpdateHash(h);
                         }
                         RemoveAction::UpdateHash(hash) => {
                             branch.set_child_hash(suffix_path.bit(0), &hash);
-                            let h = branch.hash();
+                            let h = branch.object_hash();
 
                             self.base.put(&child_path, branch);
                             return RemoveAction::UpdateHash(h);
@@ -694,7 +665,7 @@ where
                     branch.set_child(
                         prefix_path.bit(i),
                         &prefix_path.suffix(i),
-                        &prefix_data.hash(),
+                        &prefix_data.object_hash(),
                     );
                     let new_prefix = proof_path.prefix(i);
                     self.base.put(&new_prefix, branch);
@@ -720,7 +691,7 @@ where
                     new_branch.set_child(
                         prefix_path.bit(i),
                         &prefix_path.suffix(i),
-                        &branch.hash(),
+                        &branch.object_hash(),
                     );
                     new_branch.set_child(proof_path.bit(i), &proof_path.suffix(i), &hash);
                     // Saves a new branch
@@ -818,11 +789,46 @@ where
     }
 }
 
+impl<T, K, V> ObjectHash for ProofMapIndex<T, K, V>
+where
+    T: IndexAccess,
+    K: BinaryKey + ObjectHash,
+    V: BinaryValue + ObjectHash,
+{
+    /// Returns the hash of the proof map object. See [`HashTag::hash_map_node`].
+    /// For hash of the empty map see [`HashTag::empty_map_hash`].
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use exonum_merkledb::{TemporaryDB, Database, ProofMapIndex, HashTag, ObjectHash};
+    /// use exonum_crypto::Hash;
+    ///
+    /// let db = TemporaryDB::new();
+    /// let name = "name";
+    /// let fork = db.fork();
+    /// let mut index = ProofMapIndex::new(name, &fork);
+    ///
+    /// let default_hash = index.object_hash();
+    /// assert_eq!(HashTag::empty_map_hash(), default_hash);
+    ///
+    /// index.put(&default_hash, 100);
+    /// let hash = index.object_hash();
+    /// assert_ne!(hash, default_hash);
+    /// ```
+    ///
+    /// [`HashTag::hash_map_node`]: ../enum.HashTag.html#method.hash_map_node
+    /// [`HashTag::empty_map_hash`]: ../enum.HashTag.html#method.empty_map_hash
+    fn object_hash(&self) -> Hash {
+        HashTag::hash_map_node(self.merkle_root())
+    }
+}
+
 impl<'a, T, K, V> ::std::iter::IntoIterator for &'a ProofMapIndex<T, K, V>
 where
     T: IndexAccess,
     K: BinaryKey + ObjectHash,
-    V: BinaryValue + UniqueHash,
+    V: BinaryValue + ObjectHash,
 {
     type Item = (K::Owned, V);
     type IntoIter = ProofMapIndexIter<'a, K, V>;
@@ -873,7 +879,7 @@ impl<T, K, V> fmt::Debug for ProofMapIndex<T, K, V>
 where
     T: IndexAccess,
     K: BinaryKey + ObjectHash,
-    V: BinaryValue + UniqueHash + fmt::Debug,
+    V: BinaryValue + ObjectHash + fmt::Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         struct Entry<'a, T: 'a + IndexAccess, K: 'a, V: 'a + BinaryValue> {
@@ -887,7 +893,7 @@ where
         where
             T: IndexAccess,
             K: BinaryKey + ObjectHash,
-            V: BinaryValue + UniqueHash,
+            V: BinaryValue + ObjectHash,
         {
             fn new(index: &'a ProofMapIndex<T, K, V>, hash: Hash, path: ProofPath) -> Self {
                 Entry {
@@ -911,7 +917,7 @@ where
         where
             T: IndexAccess,
             K: BinaryKey + ObjectHash,
-            V: BinaryValue + UniqueHash + fmt::Debug,
+            V: BinaryValue + ObjectHash + fmt::Debug,
         {
             fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                 match self.node {