+++ src/tests.rs
@@ -16,7 +16,7 @@ use exonum_crypto::Hash;
 
 use crate::{
     Entry, Fork, KeySetIndex, ListIndex, MapIndex, ProofListIndex, ProofMapIndex, SparseListIndex,
-    ValueSetIndex
+    ValueSetIndex,
 };
 use exonum_crypto::Hash;